edit/delete timeslot, Copied property link in browser and went to signup page
- Edit timeslot check -  think it would be better to show a warning
dialogue box only when a renter's time slot is affected. What I mean is
that for e.g. a range of Sat 2pm-4pm is booked by the landlord and two
renters 2:00-2:30 and 2:30-3:00 fix the slots. If the landlord tries to
edit the time to 2pm-3pm then it should be fine but if editing the time
prompts a cancellation then we should show a dialogue. If this is too
complex then let's stick to the simple dialogue box whenever a landlord
tries to cancel.  - DONE

- Delete Timeslot - Check if te timeslot is removed but added again via
new timeslot add consider that to prevent the schedules to be cancelled
as if they belong to the newly added timeslot subslot of half an hour -
DONE

- Copied property link in browser and went to signup page, filled all
details in it and clicked signup — it didnt go to property application
page
after signup it will login directly and go to application page — send
button on this page will give error until the email is not verified